Drastic Website Additions & Changes
Added a new section below the hero where it shows users the difference of operations with and without AI+Automation system solutions. It is also crafted for mobile devices as well.

Heavily modified the services section to showcase condensed info in one spot. Also got it modified for mobile devices. Removed the custom animations for simple ones.

Improved the Our Process by aligning the icons with the text header.

Made the Calendly items' width responsive to the width of the screen.